feat(wgpu): mipmaps, anisotropic filtering, MSAA pipeline fixes   - Add WgpuMipmapGenerator: GPU render-pass blit chain for 2D and cube textures; triggered automatically when generateMipmaps=true
  - Query device maxSamplerAnisotropy at init;apply tex->anisotropy to samplers instead of hardcoding 1
  - Fix pipeline cache keys to encode sample count (bit 62 for standard,
    bit 32 for custom) so count=4 and count=1 pipelines are never aliased

  - Set effectiveSampleCount_=1 when rendering to an RT with no MSAA view

  - Fix ImGui overlay pipeline: PipelineMultisampleState.count must be 1
    since the overlay renders to the resolved non-MSAA surface texture,
    not the MSAA attachment; this was causing a wgpu validation error
    ("Incompatible sample count") on all MSAA-enabled examples with ImGui
 [skip ci]

+++ src/threepp/renderers/wgpu/WgpuPipelines.hpp
@@ -86,7 +86,9 @@ namespace threepp::wgpu {
     private:
         WgpuState& state_;
         std::unordered_map<uint64_t, PipelineEntry> pipelineCache_;
-        std::unordered_map<unsigned int, CustomPipelineEntry> customPipelineCache_;
+        // Key: (uint64_t(materialId) | (sampleCount > 1 ? (1ULL << 32) : 0))
+        // so that pipelines compiled for different MSAA counts are cached separately.
+        std::unordered_map<uint64_t, CustomPipelineEntry> customPipelineCache_;
 
         std::vector<WGPUBindGroupLayoutEntry> buildBindGroupLayoutEntries(uint64_t features) const;
